uint => uint256

+++ src/L1/FixedPriceOracle.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.13;
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
 contract FixedPriceOracle is Ownable {
-    uint public price;
+    uint256 public price;
 
-    event PriceSet(uint _newPrice);
+    event PriceSet(uint256 _newPrice);
 
     constructor(uint256 _initalPrice, address _owner) Ownable(_owner) {
         price = _initalPrice;
@@ -14,7 +14,7 @@ contract FixedPriceOracle is Ownable {
         emit PriceSet(_initalPrice);
     }
 
-    function setPrice(uint _newPrice) external onlyOwner {
+    function setPrice(uint256 _newPrice) external onlyOwner {
         price = _newPrice;
 
         emit PriceSet(_newPrice);
